refactoring so macros work with formatting
