+++ public/app/features/templating/dataMacros.ts
@@ -1,6 +1,8 @@
 import { FieldType, formattedValueToString, getDisplayProcessor, ScopedVars } from '@grafana/data';
 import { VariableCustomFormatterFn } from '@grafana/scenes';
 
+import { formatVariableValue } from './formatVariableValue';
+
 const fallbackDisplayProcessor = getDisplayProcessor();
 
 export function valueMacro(
@@ -9,6 +11,11 @@ export function valueMacro(
   fieldPath?: string,
   format?: string | VariableCustomFormatterFn
 ) {
+  const value = getValueForValueMacro(scopedVars, fieldPath);
+  return formatVariableValue(value, format);
+}
+
+function getValueForValueMacro(scopedVars?: ScopedVars, fieldPath?: string) {
   const dataContext = scopedVars?.__dataContext;
   if (!dataContext) {
     return '';
@@ -19,7 +26,7 @@ export function valueMacro(
   if (calculatedValue) {
     switch (fieldPath) {
       case 'numeric':
-        return calculatedValue.numeric;
+        return calculatedValue.numeric.toString();
       case 'raw':
         return calculatedValue.numeric;
       case 'time':
