+++ public/app/features/templating/types.ts
@@ -4,8 +4,8 @@ import { VariableCustomFormatterFn } from '@grafana/scenes';
 export interface MacroHandler {
   (
     variableName: string,
-    scopedVars?: ScopedVars,
-    fieldPath?: string,
-    format?: string | VariableCustomFormatterFn
+    scopedVars: ScopedVars | undefined,
+    fieldPath: string | undefined,
+    format: string | VariableCustomFormatterFn | undefined
   ): string;
 }
